requirements.txt
opencv-python>=4.5.5
opencv-contrib-python>=4.5.5  # For ximgproc module
numpy>=1.20.0
Pillow>=9.0.0
tqdm>=4.64.0
torch>=1.11.0
torchvision>=0.12.0
clip@git+https://github.com/openai/CLIP.git